backend/requirements.txt
# FastAPI Backend Requirements for SkillMatch AI

# Web Framework & Server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.8.0
pydantic-settings>=2.4.0

# Database
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.30
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# File handling
aiofiles>=23.0.0

# AWS/DigitalOcean Spaces
boto3>=1.34.0
botocore>=1.34.0

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.0
pdfplumber>=0.10.3

# AI/NLP - Updated for Python 3.13 compatibility
openai>=1.12.0
langchain>=0.2.0
# Skip sentence-transformers for now - has build issues with Python 3.13

# Background Tasks & Cache
celery>=5.3.0
redis>=5.0.0

# HTTP Client
httpx>=0.27.0

# Utilities
email-validator>=2.1.0
python-slugify>=8.0.0
Pillow>=10.4.0

# Logging & Monitoring
structlog>=24.0.0
sentry-sdk[fastapi]>=2.0.0

# Development & Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0

# Environment
python-dotenv>=1.0.0
